Revert "Сохранение аккаунтов ссылкой в Keychain"
This reverts commit 8a35b59e4630f1de393d38da0ba6f75547c7e658.

+++ Shared/Extension/String.swift
@@ -74,7 +74,7 @@ extension String {
         }
     }
     
-    func counterFromString() -> UInt64 {
-        UInt64(self) ?? 1
+    func counterFromString() -> Int {
+        Int(self) ?? 1
     }
 }
